chore: split two files over the 800-line soft cap into submodules
db/src/metadata_overrides/upsert.rs (~825 lines) becomes upsert/{mod,merge,cover}.rs:
mod.rs keeps the error enum, shared row-write/touch/cache-invalidate
helpers, single-book upsert, and plain get/delete/bulk-load; merge.rs
holds the single-book and bulk merge paths; cover.rs holds the
cover-override file I/O plus the overrides-to-EbookMetadata read merge.
Everything stays re-exported at the `upsert::` path so no caller outside
the module changes.

frontend/src/pages/settings/users.rs (~812 lines) becomes
users/{mod,modals,registration}.rs: mod.rs keeps the table shell, row,
and permission components; modals.rs holds the New/Edit/Delete modals;
registration.rs holds the self-registration switch and its pure
load/toggle-guard helpers, with its own sibling tests.rs (moved with
the code it tests, per rule 03).

Pure mechanical extraction — no behavior change. Closes #1709.
